try the new tools from mvnparent; unchanged copy for the first time

+++ release/ckdep.sh
@@ -24,24 +24,62 @@
 # initialisation
 export LC_ALL=C
 unset LANGUAGE
-PS4='++ '
 set -e
 set -o pipefail
-parentpompath=..
-mvnprofiles=''
-cd "$(dirname "$0")"
-saveIFS=$' \t\n'
-print -ru2 -- '[INFO] ckdep.sh starting'
-abend=0
+exec 8>&1 9>&2
 
+function errmsg {
+	set -o noglob
+	local x
+
+	print -ru9 -- "$1"
+	shift
+	IFS=$'\n'
+	set -- $*
+	for x in "$@"; do
+		print -ru9 -- "| $x"
+	done
+}
+function die {
+	errmsg "$@"
+	exit 1
+}
+
+# check that we’re really run from mvn
+[[ -n $CKDEP_RUN_FROM_MAVEN ]] || {
+	(
+		(
+			CKDEP_RUN_FROM_MAVEN=no "$0" "$@" | \
+			    sed 's/^/[INFO] /g'
+		) 2>&1 >&3 | \
+		    sed 's/^/[ERROR] /g'
+	) 3>&1
+	exit $?
+}
+# initialisation
+cd "$(dirname "$0")"
+ancillarypath=$PWD
+. ./cksrc.sh
+
+set +e
 x=$(sed --posix 's/u\+/x/g' <<<'fubar fuu' 2>&1) && alias 'sed=sed --posix'
 x=$(sed -e 's/u\+/x/g' -e 's/u/X/' <<<'fubar fuu' 2>&1)
 case $?:$x {
 (0:fXbar\ fuu) ;;
-(*)
-	print -ru2 -- '[ERROR] your sed is not POSIX compliant'
-	exit 1 ;;
+(*) die 'your sed is not POSIX compliant' ;;
 }
+set -e
+
+function logline {
+	# tee /dev/fd/$1 # not reliable
+	while IFS= read -r line; do
+		print -r -- "$line"
+		print -ru8 -- "${line/#'[INFO] '/\| }"
+	done
+}
+
+print -ru8 -- 'ckdep.sh starting'
+abend=0
 
 # get project metadata
 <"$parentpompath/pom.xml" xmlstarlet sel \
@@ -50,13 +88,18 @@ case $?:$x {
     -c /pom:project/pom:artifactId -n \
     -c /pom:project/pom:version -n \
     |&
-IFS= read -pr pgID
-IFS= read -pr paID
-IFS= read -pr pVSN
+e=0
+IFS= read -pr pgID || e=1
+IFS= read -pr paID || e=1
+IFS= read -pr pVSN || e=1
+[[ $e = 0 && -n $pgID && -n $paID && -n $pVSN ]] || die \
+    'could not get project metadata' \
+    "pgID=$pgID" "paID=$paID" "pVSN=$pVSN"
+
 # check old file is sorted
 sort -uo ckdep.tmp ckdep.lst
 if ! cmp -s ckdep.lst ckdep.tmp; then
-	print -ru2 -- '[WARNING] list of dependencies was not sorted!'
+	errmsg 'list of dependencies was not sorted!'
 	cat ckdep.tmp >ckdep.lst
 	(( abend |= 1 ))
 fi
@@ -64,17 +107,21 @@ fi
 function domvn {
 	mvn -B "$@" dependency:list 2>&1 | scanmvn
 }
+set -A scanmvn_excludes -- \
+    -e '/^\[INFO]    '$pgID:$paID':/d'
+for x in "${depexcludes[@]}"; do
+	scanmvn_excludes+=(-e '/^\[INFO]    '"$x/d")
+done
 function scanmvn {
-	tee /dev/stderr | sed -n \
+	logline | sed -n \
 	    -e 's/ -- module .*$//' \
-	    -e '/^\[INFO]    org.evolvis.veraweb:/d' \
-	    -e '/^\[INFO]    org.evolvis.veraweb.middleware:/d' \
+	    "${scanmvn_excludes[@]}" \
 	    -e '/^\[INFO]    \([^:]*\):\([^:]*\):jar:\([^:]*\):\([^:]*\)$/s//\1:\2 \3 \4/p'
 }
 function doscopes {
 	local lastgav lastscope ga v scope rest
 
-	sort -u | while read ga v scope rest; do
+	while read -r ga v scope rest; do
 		# compile scope supersets provided scope, either supersets test
 		case "$lastgav $lastscope:$scope" {
 		("$ga:$v "compile:provided|"$ga:$v "@(compile|provided):test) ;;
@@ -83,7 +130,8 @@ function doscopes {
 		lastgav=$ga:$v lastscope=$scope
 	done
 }
-(cd "$parentpompath" && domvn $mvnprofiles) | doscopes | sort -u >ckdep.mvn.tmp
+{ cd "$parentpompath" && domvn $mvnprofiles; } | \
+    sort -u | doscopes | sort -u >ckdep.mvn.tmp
 # deal with embedded copies
 function dopom {
 	local scope=$1; shift
@@ -105,7 +153,7 @@ function dopom {
 	for gav in "$@"; do
 		IFS=:
 		set -- $gav
-		IFS=$saveIFS
+		IFS=$' \t\n'
 		cat <<-EOF
 			<dependency>
 				<groupId>$1</groupId>
@@ -125,26 +173,26 @@ function dopom {
 }
 function dodoc_! {
 	local scope=$1; shift
-	print -ru2 -- "[INFO] resolving embedded-code-copy-document-$1 [$scope]"
+	print -ru8 -- "resolving embedded-code-copy-document-$1 [$scope]"
 	shift
 	set -o noglob
 	for gav in "$@"; do
 		IFS=:
 		set -- $gav
-		IFS=$saveIFS
+		IFS=$' \t\n'
 		print -r -- "[INFO]    $1:$2:jar:$3:doc-only"
 	done | scanmvn >>ckdep.pom.tmp
 	set +o noglob
 }
 function dodoc_+ {
 	local scope=$1; shift
-	print -ru2 -- "[INFO] resolving embedded-code-copy-insert-$1 [$scope]"
+	print -ru8 -- "resolving embedded-code-copy-insert-$1 [$scope]"
 	shift
 	set -o noglob
 	for gav in "$@"; do
 		IFS=:
 		set -- $gav
-		IFS=$saveIFS
+		IFS=$' \t\n'
 		print -r -- "[INFO]    $1:$2:jar:$3:unreleased"
 	done | scanmvn >>ckdep.pom.tmp
 	set +o noglob
@@ -154,7 +202,7 @@ set -A cc_where
 set -A cc_which
 set -A cc_nomvn
 ncc=0
-[[ ! -s ckdep.ins ]] || while read first rest; do
+[[ ! -s ckdep.ins ]] || while read -r first rest; do
 	[[ $first != ?('#'*) ]] || continue
 	if [[ $first = [+!]* ]]; then
 		cc_nomvn[ncc]=${first::1}
@@ -173,18 +221,16 @@ function depround {
 		if [[ ${cc_where[i]} = "$rest" ]]; then
 			# insert embedded dependencies
 			if [[ -n ${cc_found[i]} ]]; then
-				print -ru2 -- "[ERROR]" matched \
-				    ${cc_where[i]} ${cc_which[i]} \
-				    multiple times
+				errmsg "matched ${cc_where[i]} ${cc_which[i]} multiple times"
 				(( abend |= 2 ))
 			fi
 			cc_found[i]=x
 			if [[ -n ${cc_nomvn[i]} ]]; then
-				print -ru2 -- "[INFO]" documentation-only \
+				print -ru8 -- documentation-only \
 				    dependencies for ${cc_where[i]}
 				dodoc_"${cc_nomvn[i]}" $scope $i ${cc_which[i]}
 			else
-				print -ru2 -- "[INFO]" analysing embedded \
+				print -ru8 -- analysing embedded \
 				    code copies found inside ${cc_where[i]}
 				dopom $scope $i ${cc_which[i]}
 			fi
@@ -194,17 +240,16 @@ function depround {
 		fi
 	done
 	if [[ -n $vf ]]; then
-		print -ru2 -- "[ERROR]" version mismatch: $vf
+		errmsg "version mismatch: $vf"
 		(( abend |= 2 ))
 	fi
 	if [[ -s ckdep.pom.tmp ]]; then
 		local recurse
 		set -A recurse
-		doscopes <ckdep.pom.tmp |&
-		while read -p ga v x; do
+		<ckdep.pom.tmp sort -u | doscopes |&
+		while read -pr ga v x; do
 			if [[ $x != @(doc-only|unreleased|"$scope") ]]; then
-				print -ru2 -- "[ERROR]" unexpected scope \
-				    $ga $v "${x@Q}" for $rest $scope
+				errmsg "unexpected scope $ga $v ${x@Q} for $rest $scope"
 				(( abend |= 2 ))
 			fi
 			[[ $x = doc-only ]] || print -ru4 -- $ga $v $scope
@@ -216,7 +261,7 @@ function depround {
 		done
 	fi
 }
-while read first v scope; do
+while read -r first v scope; do
 	print -ru4 -- $first $v $scope
 	print -ru5 -- $first $v $scope ok
 	depround $first $v $scope
@@ -224,20 +269,19 @@ done <ckdep.mvn.tmp 4>ckdep.mvp.tmp 5>ckdep.audit.tmp
 i=-1
 while (( ++i < ncc )); do
 	if [[ -z ${cc_found[i]} ]]; then
-		print -ru2 -- "[ERROR]" did not match \
-		    ${cc_where[i]} ${cc_which[i]}
+		errmsg "did not match ${cc_where[i]} ${cc_which[i]}"
 		(( abend |= 2 ))
 	fi
 done
 # ship source only for some scopes
-doscopes <ckdep.mvp.tmp | while read ga v scope rest; do
+<ckdep.mvp.tmp sort -u | doscopes | while read -r ga v scope rest; do
 	[[ $scope != @(compile|runtime) ]] || print -r -- ${ga/:/ } $v
 done | sort -u >ckdep.mvn.tmp
 # add static dependencies from embedded files, for SecurityWatch
 [[ ! -s ckdep.inc ]] || cat ckdep.inc >>ckdep.audit.tmp
 # generate file with changed dependencies set to be a to-do item
 # except we don’t licence-analyse test-only dependencies
-doscopes <ckdep.audit.tmp | sort -u >ckdep.tmp
+<ckdep.audit.tmp sort -u | doscopes | sort -u >ckdep.tmp
 {
 	comm -13 ckdep.lst ckdep.tmp | sed 's/ ok$/ TO''DO/'
 	comm -12 ckdep.lst ckdep.tmp
@@ -245,28 +289,28 @@ doscopes <ckdep.audit.tmp | sort -u >ckdep.tmp
 
 # check if the list changed
 if cmp -s ckdep.lst ckdep.tmp && cmp -s ckdep.mvn ckdep.mvn.tmp; then
-	print -ru2 -- '[INFO] list of dependencies did not change'
+	print -ru8 -- 'list of dependencies did not change'
 else
 	(diff -u ckdep.lst ckdep.tmp || :)
 	# make the new list active
 	mv -f ckdep.mvn.tmp ckdep.mvn
 	mv -f ckdep.tmp ckdep.lst
 	# inform the user
-	print -ru2 -- '[WARNING] list of dependencies changed!'
+	errmsg 'list of dependencies changed!'
 	(( abend |= 1 ))
 fi
 rm -f ckdep.pom ckdep.tmp ckdep.*.tmp
 # check if anything needs to be committed
 if (( abend & 1 )); then
-	print -ru2 -- '[ERROR] please commit the changed ckdep.{lst,mvn} files!'
+	errmsg 'please commit the changed ckdep.{lst,mvn} files!'
 fi
 
 # fail a release build if dependency licence review has a to-do item
 [[ $IS_M2RELEASEBUILD != true ]] || \
     if grep -e ' TO''DO$' -e ' FA''IL$' ckdep.lst; then
-	print -ru2 -- '[ERROR] licence review incomplete'
+	errmsg 'licence review incomplete'
 	(( abend |= 4 ))
 fi
 
-print -ru2 -- '[INFO] ckdep.sh finished with errorlevel' $((#abend))
+print -ru8 -- 'ckdep.sh finished with errorlevel' $((#abend))
 exit $abend
